+++ WeatherDataProvider2.py
@@ -37,7 +37,32 @@ def loadLatestWeatherData(config):
     url = 'https://api.forecast.solar/estimate/{}/{}/{}/{}/{}'.format(lat, lon, dec, az, kwp)
     try:
         apiResponse = requests.get(url, timeout=12.50)
-        return dict(json.loads(apiResponse.text))
+        json_data1 = dict(json.loads(apiResponse.text))
+        # Hier werden fuer ein evtl. zweites Feld mit anderer Ausrichtung die Prognosewerte eingearbeitet
+        if config['pv.strings']['anzahl'] == '2':
+            dict_watts = {}
+            dict_watt_hours = {}
+            lat = config['forecast.solar2']['lat']
+            lon = config['forecast.solar2']['lon']
+            dec = config['forecast.solar2']['dec']
+            az = config['forecast.solar2']['az']
+            kwp = config['forecast.solar2']['kwp']
+
+            url = 'https://api.forecast.solar/estimate/{}/{}/{}/{}/{}'.format(lat, lon, dec, az, kwp)
+            apiResponse2 = requests.get(url, timeout=12.50)
+            json_data2 = dict(json.loads(apiResponse2.text))
+		
+            for key, value in json_data1.get('result',{}).get('watts',{}).items():
+                dict_watts[key]=value
+            for key, value in json_data2.get('result',{}).get('watts',{}).items():
+                dict_watts[key]=dict_watts[key]+value
+            for key, value in json_data1.get('result',{}).get('watt_hours',{}).items():
+                dict_watt_hours[key]=value
+            for key, value in json_data2.get('result',{}).get('watt_hours',{}).items():
+                dict_watt_hours[key]=dict_watt_hours[key]+value
+            json_data1['result']['watts']=dict_watts
+            json_data1['result']['watt_hours']=dict_watt_hours
+        return(json_data1)
     except:
         return("False")